Player Package Refactor
