+++ .gitignore
@@ -3,3 +3,5 @@
 /GameResources/__pycache__/
 /Experiments/
 /GameReplays/
+Players/__pycache__/AlgorithmicPlayers.cpython-311.pyc
+Players/__pycache__/SimplePlayers.cpython-311.pyc
